requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
greenlet==2.0.1
idna==3.4
multidict==6.0.4
psycopg2==2.9.5
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.1
requests==2.28.1
SQLAlchemy==1.4.45
urllib3==1.26.13
yarl==1.8.2
